LayoutTests/fast/forms/input-minmax-expected.txt

Tests the behavior of .min and .max of HTMLInputElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.min is ""
PASS input.max is ""
PASS input.min is "0"
PASS input.max is "100"
PASS input.min is "abc"
PASS input.max is "qwerty"
PASS input.getAttribute("min") is "foo"
PASS input.getAttribute("max") is "bar"
PASS input.getAttribute("min") is ""
PASS input.getAttribute("max") is ""
PASS input.min is ""
PASS input.max is ""
PASS successfullyParsed is true

TEST COMPLETE
